requirements.txt
pandas==2.2.1
numpy==1.26.4
xgboost==2.0.3
notebook==7.1.2
seaborn==0.13.2
Flask==3.0.2
mlflow==2.11.3
python-box==7.1.1
PyYAML==6.0.1
gdown==5.1.0
scikit-learn==1.4.1post1
-e .